--- client/src/store/banner.ts
@@ -1,5 +1,0 @@
-import { atomWithLocalStorage } from './utils';
-
-const hideBannerHint = atomWithLocalStorage('hideBannerHint', [] as string[]);
-
-export default { hideBannerHint };
