+++ client/src/store/families.ts
@@ -75,16 +75,16 @@ const conversationByIndex = atomFamily<TConversation | null, string | number>({
       onSet(async (newValue) => {
         const index = Number(node.key.split('__')[1]);
         logger.log('conversation', 'Setting conversation:', { index, newValue });
-        if (newValue?.assistant_id) {
+        if (newValue?.assistant_id != null && newValue.assistant_id) {
           localStorage.setItem(
             `${LocalStorageKeys.ASST_ID_PREFIX}${index}${newValue.endpoint}`,
             newValue.assistant_id,
           );
         }
-        if (newValue?.agent_id) {
+        if (newValue?.agent_id != null && newValue.agent_id) {
           localStorage.setItem(`${LocalStorageKeys.AGENT_ID_PREFIX}${index}`, newValue.agent_id);
         }
-        if (newValue?.spec) {
+        if (newValue?.spec != null && newValue.spec) {
           localStorage.setItem(LocalStorageKeys.LAST_SPEC, newValue.spec);
         }
         if (newValue?.tools && Array.isArray(newValue.tools)) {
@@ -238,6 +238,11 @@ const audioRunFamily = atomFamily<string | null, string | number | null>({
   default: null,
 });
 
+const messagesSiblingIdxFamily = atomFamily<number, string | null | undefined>({
+  key: 'messagesSiblingIdx',
+  default: 0,
+});
+
 function useCreateConversationAtom(key: string | number) {
   const [keys, setKeys] = useRecoilState(conversationKeysAtom);
   const setConversation = useSetRecoilState(conversationByIndex(key));
@@ -260,7 +265,7 @@ function useClearConvoState() {
         const conversationKeys = await snapshot.getPromise(conversationKeysAtom);
 
         for (const conversationKey of conversationKeys) {
-          if (skipFirst && conversationKey == 0) {
+          if (skipFirst === true && conversationKey == 0) {
             continue;
           }
 
@@ -298,7 +303,7 @@ function useClearSubmissionState() {
         logger.log('submissionKeys', submissionKeys);
 
         for (const key of submissionKeys) {
-          if (skipFirst && key == 0) {
+          if (skipFirst === true && key == 0) {
             continue;
           }
 
@@ -320,12 +325,12 @@ function useClearLatestMessages(context?: string) {
       async (skipFirst?: boolean) => {
         const latestMessageKeys = await snapshot.getPromise(latestMessageKeysSelector);
         logger.log('[clearAllLatestMessages] latestMessageKeys', latestMessageKeys);
-        if (context) {
+        if (context != null && context) {
           logger.log(`[clearAllLatestMessages] context: ${context}`);
         }
 
         for (const key of latestMessageKeys) {
-          if (skipFirst && key == 0) {
+          if (skipFirst === true && key == 0) {
             continue;
           }
 
@@ -380,6 +385,7 @@ export default {
   showBingToneSettingFamily,
   showPopoverFamily,
   latestMessageFamily,
+  messagesSiblingIdxFamily,
   allConversationsSelector,
   conversationByKeySelector,
   useClearConvoState,
